requirements.txt
google-genai
yt-dlp
pydantic
python-dotenv
fastapi
uvicorn
httpx
gspread
google-auth
